Persistencia login

+++ UI_Escritorio/FormPlanes.cs
@@ -31,20 +31,28 @@ namespace UI_Escritorio
 
         private async void btnBorrar_Click(object sender, EventArgs e)
         {
-            int id = this.SelectedItem().IdPlan;
-            await PlanesApi.DeleteAsync(id);
-            this.GetAllAndLoad();
+            Plan planSelected = this.SelectedItem();
+            if(planSelected != null)
+            { 
+                int id = planSelected.IdPlan;
+                await PlanesApi.DeleteAsync(id);
+                this.GetAllAndLoad();
+            }
         }
 
         private async void btnEditar_Click(object sender, EventArgs e)
         {
-            FormPlanesDetalle planDetalle = new FormPlanesDetalle();
-            int id = this.SelectedItem().IdPlan;
-            Plan plan = await PlanesApi.GetAsync(id);
-            planDetalle.EditMode = true;
-            planDetalle.Plan = plan;
-            planDetalle.ShowDialog();
-            this.GetAllAndLoad();
+            Plan planSelected = this.SelectedItem();
+            if (planSelected != null)
+            {
+                FormPlanesDetalle planDetalle = new FormPlanesDetalle();
+                int id = planSelected.IdPlan;
+                Plan plan = await PlanesApi.GetAsync(id);
+                planDetalle.EditMode = true;
+                planDetalle.Plan = plan;
+                planDetalle.ShowDialog();
+                this.GetAllAndLoad();
+            }
         }
 
         private void btnNuevo_Click(object sender, EventArgs e)
@@ -86,7 +94,18 @@ namespace UI_Escritorio
 
         private Plan SelectedItem()
         {
-            return (Plan)dgvPlanes.SelectedRows[0].DataBoundItem;
+
+            if (dgvPlanes.SelectedRows.Count > 0)
+            {
+                Plan plan = (Plan)dgvPlanes.SelectedRows[0].DataBoundItem;
+                if (plan != null)
+                {
+                    return plan;
+                }
+            }
+
+            MessageBox.Show("Debes listar y seleccionar un plan", "Planes", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            return null;
         }
 
     }
